frontend/requirements.txt
streamlit
requests
pandas
numpy
datetime
matplotlib
PyJWT